Re-organize test driver code
