+++ test/ui/driver/read-all.js
@@ -1,0 +1,15 @@
+'use strict';
+var Promise = require('bluebird');
+
+/**
+ * Get the visible text in one or more LeadFoot Element instances.
+ *
+ * @param {Array<Leadfoot.Element>}
+ *
+ * @returns {Promise} Eventual Array of strings
+ */
+module.exports = function(els) {
+  return Promise.all(els.map(function(el) {
+    return el.getVisibleText();
+  }));
+};
